Moved LongestPrefix query to 'query/comparative' package.
